✨ Allow using a fixture builder function in stubbed endpoints
When stubbing an endpoint, we can now provide a fixture builder function (in place of a fixture object).

The function takes the HTTP request as parameter, so that we can return a parametrized response (for example based on a query parameter or on the payload's content).

The fixture builder function can be overridden (using `RouteConfig.withOverride()`) or mapped (using `RouteConfig.mappingFixture`).